fix: the wrong message would be flashed in the saved messages view
The bug was due to this screen displaying messages
in an order different from the original list,
which is sorted by lastUsed.
To solve it, I moved the position <-> message
correspondence from the fragments displaying
message lists to the MessageListAdapter itself,
which already had a message list anyway...
This also has the added benefits of factorising
this logic and moving it away from a place it
didn't really belong.